+++ src/transformnd/constants.py
@@ -1,1 +1,0 @@
-UNSPECIFIED_SPACE_NAME = "???"